Complete initialization
