+++ tests/GraphCycleError.py
@@ -1,21 +1,15 @@
-def check(vertex_ids, edge_ids,edge_enabled,edge_vertex_id_pairs):
-    size=len(vertex_ids)
-    sparseMatrix= [ [ 0 for i in range(size) ] for j in range(size) ]
-    for i in range(size):
-        for j in range(size):
-            if(((vertex_ids[i],vertex_ids[j]) in edge_vertex_id_pairs) and sparseMatrix[i][j]==0):
-                if edge_enabled[edge_vertex_id_pairs.index((vertex_ids[i],vertex_ids[j]))]:
-                    sparseMatrix[i][j]=vertex_ids[j]
-                    sparseMatrix[j][i]=vertex_ids[i]
-            elif(((vertex_ids[j],vertex_ids[i]) in edge_vertex_id_pairs)  and sparseMatrix[i][j]==0):
-                if (edge_enabled[edge_vertex_id_pairs.index((vertex_ids[j],vertex_ids[i]))]==True):
-                    sparseMatrix[i][j]=vertex_ids[j]
-                    sparseMatrix[j][i]=vertex_ids[i]
-    print(sparseMatrix)
+import networkx as nx
+def check(edge_enabled,edge_vertex_id_pairs):
+    G = nx.Graph()
+    for (u, v), enabled in zip(edge_vertex_id_pairs, edge_enabled):
+        if enabled:
+            G.add_edge(u, v)
 
-edge_ids = [1, 3, 5, 7, 8, 9]
-vertex_ids = [0, 2, 4, 6, 10]
-edge_enabled = [True, True, True, False, False, True]
-edge_vertex_id_pairs = [(0, 2), (0, 4), (0, 6), (2, 4), (2, 10), (4, 6)]
+    has_cycle = nx.is_forest(G)
+    if (has_cycle==False):
+        raise Exception("GraphCycleError")
+    
+#edge_enabled = [True, True, True, False, True, False]
+#edge_vertex_id_pairs = [(0, 2), (0, 4), (0, 6), (2, 4), (2, 10), (4, 6)]
 
-check(vertex_ids, edge_ids,edge_enabled,edge_vertex_id_pairs)
+#check(edge_enabled,edge_vertex_id_pairs)
