+++ tests/input_dataset.py
@@ -6,6 +6,8 @@ import InputLengthDoesNotMatchError
 import networkx as nx
 import numpy as np
 import pandas as pd
+import GraphCycleError
+import GraphNotFullyConnectedError
 
 """
 vertex_ids: list of vertex ids
@@ -30,8 +32,8 @@ class GraphProcessor:
         IDNotFoundError.check_pairs(edge_vertex_id_pairs, vertex_ids)
         InputLengthDoesNotMatchError.check_enabled(edge_enabled, edge_ids)
         IDNotFoundError.check_source(source_vertex_id, vertex_ids)
-
-        
+        GraphNotFullyConnectedError.check(vertex_ids, edge_ids,edge_enabled,edge_vertex_id_pairs)
+        GraphCycleError.check(edge_enabled,edge_vertex_id_pairs)
 
         self.vertex_ids = vertex_ids
         self.edge_ids = edge_ids
@@ -43,7 +45,7 @@ class GraphProcessor:
 vertex_ids = [0, 2, 4, 6, 10]
 edge_ids = [1, 3, 5, 7, 8, 9]
 edge_vertex_id_pairs = [(0, 2), (0, 4), (0, 6), (2, 4), (2, 10), (4, 6)]
-edge_enabled = [True, True, True, False, False, True]
+edge_enabled = [True, True, True, False, True, True]
 source_vertex_id = 0
 
 g = GraphProcessor(vertex_ids, edge_ids, edge_vertex_id_pairs, edge_enabled, source_vertex_id)
